Add files via upload
In my final year project, I was involved in the development of a 5G SDN Testbed that aims to defend against DDos attacks. This file is a script written in python that aimed to set up a Vlan host as well as integrate with the 5G testbed that was developed. There is also a start-up guide attached, and the aim of this document is to assist future members of the project to be able to import the VMs created without too much difficulty.